view_job_records(): move into own module
Move view_job_records() and its helper functions
into its own file, job_archive_interface.py.

Fixes #56
